Split components into individual files
